--- server/models/Posts.js
@@ -1,45 +1,0 @@
-const mongoose = require('mongoose');
-
-const { Schema } = mongoose;
-
-const postSchema = new Schema({
-    jobsNeeded: [
-        {
-            name: {
-                type: String,
-                required: true,
-            },
-            jobCustomer: {
-                type: String,
-                required: true,
-            },
-            description: {
-                type: String,
-                required: true,
-                minLength: 1,
-                maxLength: 1000,
-                trim: true,
-            },
-            skills: {
-
-                required: true,
-            },
-            city: {
-                type: String,
-                required: true,
-            },
-            needDate: {
-                type: Date,
-                required: false,
-            },
-            createdDate: {
-                type: Date,
-                default: Date.now,
-            },
-        },
-    ],
-});
-
-const Posts = mongoose.model('Posts', postSchema);
-
-module.exports = Posts;
