+++ server/models/index.js
@@ -1,7 +1,7 @@
 const User = require('./User');
 const Contractor = require('./Contractor');
 const Customer = require('./Customer');
-const Posts = require('./Posts');
+const Job = require('./Job');
 const Proposal = require('./Proposal');
 
 module.exports = { User, Contractor, Customer, Posts, Proposal };
